+++ userspace/lib/src/sys/mman.c
@@ -22,11 +22,12 @@
 // along with VesperaOS. If not, see <https://www.gnu.org/licenses/>.
 
 #include <stddef.h>
+#include <stdio.h>
 #include <stdlib.h>
-#include <sysstd.h>
 #include <sys/mman.h>
+#include <sysstd.h>
 
-void* mmap(void* addr, size_t length, int prot, int flags, int handle, size_t offset) {
+void* mmap(void* addr, size_t length, uint64_t prot, uint64_t flags, uint64_t handle, size_t offset) {
     int64_t ret = sys_mmap((uint64_t)addr, length, prot, flags, handle, offset);
     if (ret < 0) {
         errno = -ret;
